explain/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "explain-agent-bearing"
version = "0.1.0"
dependencies = [
  "fastapi>=0.95",
  "uvicorn[standard]>=0.22",
  "python-dotenv",
  "langchain-openai"
]
